+++ routes/actions.php
@@ -212,13 +212,13 @@ Route::prefix(implode('/', [
         Route::middleware(RequireAdminChanges::class)->post('addresses/save-field-layout', [AddressesController::class, 'saveFieldLayout']);
 
         // App
-        Route::any('app/get-cp-alerts', [CpAlertsController::class, 'index']);
-        Route::any('app/shun-cp-alert', [CpAlertsController::class, 'destroy']);
-        Route::any('app/set-license-shun-cookie', [LicensesController::class, 'setShunCookie']);
+        Route::post('app/get-cp-alerts', [CpAlertsController::class, 'index']);
+        Route::post('app/shun-cp-alert', [CpAlertsController::class, 'destroy']);
+        Route::post('app/set-license-shun-cookie', [LicensesController::class, 'setShunCookie']);
         Route::middleware(RequireAdmin::class)->post('app/get-plugin-license-info', [CraftCms\Cms\Http\Controllers\App\PluginsController::class, 'getLicenseInfo']);
         Route::middleware(RequireAdminChanges::class)->post('app/update-plugin-license', [CraftCms\Cms\Http\Controllers\App\PluginsController::class, 'updateLicense']);
-        Route::any('app/render-elements', [RenderController::class, 'elements']);
-        Route::any('app/render-components', [RenderController::class, 'components']);
+        Route::post('app/render-elements', [RenderController::class, 'elements']);
+        Route::post('app/render-components', [RenderController::class, 'components']);
 
         // Auth methods
         Route::post('auth/method-setup-html', [AuthMethodController::class, 'setupHtml']);
@@ -393,15 +393,15 @@ Route::prefix(implode('/', [
         Route::any('preview/create-token', [PreviewController::class, 'createToken']);
 
         // Queue
-        Route::any('queue/run', [QueueController::class, 'run']);
-        Route::any('queue/get-job-info', [QueueController::class, 'jobInfo']);
-        Route::any('queue/release', [QueueController::class, 'cancel']);
-        Route::any('queue/release-all', [QueueController::class, 'cancelAll']);
-        Route::any('queue/retry', [QueueController::class, 'retry']);
-        Route::any('queue/retry-all', [QueueController::class, 'retryAll']);
+        Route::post('queue/run', [QueueController::class, 'run']);
+        Route::get('queue/get-job-info', [QueueController::class, 'jobInfo']);
+        Route::post('queue/release', [QueueController::class, 'cancel']);
+        Route::post('queue/release-all', [QueueController::class, 'cancelAll']);
+        Route::post('queue/retry', [QueueController::class, 'retry']);
+        Route::post('queue/retry-all', [QueueController::class, 'retryAll']);
 
         // Relational fields
-        Route::any('relational-fields/structured-input-html', [RelationalFieldsController::class, 'structuredInputHtml']);
+        Route::post('relational-fields/structured-input-html', [RelationalFieldsController::class, 'structuredInputHtml']);
 
         // Widgets
         Route::post('dashboard/create-widget', [WidgetsController::class, 'store']);