+++ CHANGELOG.md
@@ -43,6 +43,12 @@
 - Fixed a `yii\base\InvalidConfigException` that occurred when a Yii2-based plugin registered an asset bundle. ([#18818](https://github.com/craftcms/cms/issues/18818))
 - Fixed a bug where using `{{ successMessageInput() }}` would not decrypt the resulting message for the flash message.
 - Fixed a bug where a missing widget from an uninstalled plugin would throw instead of mapping to a MissingWidget.
+- Fixed a bug where an address' ownership ids could be overridden unintentionally.
+- Fixed a bug where `getHasSsoIdentity()` would return `false` when Socialite was not installed but the user had an SSO identity.
+- Fixed a bug where the site's offline status was not being enforced on matched element routes.
+- Fixed a user photo validation issue with file extensions.
+- Fixed a bug where legacy controllers could return `null` but were not considered handled.
+- Improved performance of the dashboard by reducing the amount of queries for widgets
 
 ## 6.0.0-alpha.1 - 2026-05-06